Update models after rebase
