+++ src/api/entities/sys/TypeRefs.ts
@@ -2698,105 +2698,6 @@ export type UpgradePriceServiceReturn = {
 	teamsBusinessPrices: PlanPrices;
 	teamsPrices: PlanPrices;
 }
-export const UsageTestAssignmentTypeRef: TypeRef<UsageTestAssignment> = new TypeRef("sys", "UsageTestAssignment")
-
-export function createUsageTestAssignment(values?: Partial<UsageTestAssignment>): UsageTestAssignment {
-	return Object.assign(create(typeModels.UsageTestAssignment, UsageTestAssignmentTypeRef), values)
-}
-
-export type UsageTestAssignment = {
-	_type: TypeRef<UsageTestAssignment>;
-
-	_id: Id;
-	name: string;
-	state: NumberString;
-	testId: Id;
-	variant: null | NumberString;
-
-	stages: UsageTestStage[];
-}
-export const UsageTestAssignmentInTypeRef: TypeRef<UsageTestAssignmentIn> = new TypeRef("sys", "UsageTestAssignmentIn")
-
-export function createUsageTestAssignmentIn(values?: Partial<UsageTestAssignmentIn>): UsageTestAssignmentIn {
-	return Object.assign(create(typeModels.UsageTestAssignmentIn, UsageTestAssignmentInTypeRef), values)
-}
-
-export type UsageTestAssignmentIn = {
-	_type: TypeRef<UsageTestAssignmentIn>;
-
-	_format: NumberString;
-	testDeviceId: null | Id;
-}
-export const UsageTestAssignmentOutTypeRef: TypeRef<UsageTestAssignmentOut> = new TypeRef("sys", "UsageTestAssignmentOut")
-
-export function createUsageTestAssignmentOut(values?: Partial<UsageTestAssignmentOut>): UsageTestAssignmentOut {
-	return Object.assign(create(typeModels.UsageTestAssignmentOut, UsageTestAssignmentOutTypeRef), values)
-}
-
-export type UsageTestAssignmentOut = {
-	_type: TypeRef<UsageTestAssignmentOut>;
-
-	_format: NumberString;
-	testDeviceId: Id;
-
-	assignments: UsageTestAssignment[];
-}
-export const UsageTestMetricConfigTypeRef: TypeRef<UsageTestMetricConfig> = new TypeRef("sys", "UsageTestMetricConfig")
-
-export function createUsageTestMetricConfig(values?: Partial<UsageTestMetricConfig>): UsageTestMetricConfig {
-	return Object.assign(create(typeModels.UsageTestMetricConfig, UsageTestMetricConfigTypeRef), values)
-}
-
-export type UsageTestMetricConfig = {
-	_type: TypeRef<UsageTestMetricConfig>;
-
-	_id: Id;
-	name: string;
-	type: NumberString;
-}
-export const UsageTestMetricDataTypeRef: TypeRef<UsageTestMetricData> = new TypeRef("sys", "UsageTestMetricData")
-
-export function createUsageTestMetricData(values?: Partial<UsageTestMetricData>): UsageTestMetricData {
-	return Object.assign(create(typeModels.UsageTestMetricData, UsageTestMetricDataTypeRef), values)
-}
-
-export type UsageTestMetricData = {
-	_type: TypeRef<UsageTestMetricData>;
-
-	_id: Id;
-	name: string;
-	value: string;
-}
-export const UsageTestParticipationInTypeRef: TypeRef<UsageTestParticipationIn> = new TypeRef("sys", "UsageTestParticipationIn")
-
-export function createUsageTestParticipationIn(values?: Partial<UsageTestParticipationIn>): UsageTestParticipationIn {
-	return Object.assign(create(typeModels.UsageTestParticipationIn, UsageTestParticipationInTypeRef), values)
-}
-
-export type UsageTestParticipationIn = {
-	_type: TypeRef<UsageTestParticipationIn>;
-
-	_format: NumberString;
-	stage: NumberString;
-	testDeviceId: Id;
-	testId: Id;
-
-	metrics: UsageTestMetricData[];
-}
-export const UsageTestStageTypeRef: TypeRef<UsageTestStage> = new TypeRef("sys", "UsageTestStage")
-
-export function createUsageTestStage(values?: Partial<UsageTestStage>): UsageTestStage {
-	return Object.assign(create(typeModels.UsageTestStage, UsageTestStageTypeRef), values)
-}
-
-export type UsageTestStage = {
-	_type: TypeRef<UsageTestStage>;
-
-	_id: Id;
-	name: string;
-
-	metrics: UsageTestMetricConfig[];
-}
 export const UserTypeRef: TypeRef<User> = new TypeRef("sys", "User")
 
 export function createUser(values?: Partial<User>): User {
