--- src/api/entities/sys/sysModelMap.ts
@@ -1,200 +1,0 @@
-const map = {
-    AccountingInfo: () => import('./AccountingInfo.js'),
-    AdministratedGroup: () => import('./AdministratedGroup.js'),
-    AdministratedGroupsRef: () => import('./AdministratedGroupsRef.js'),
-    AlarmInfo: () => import('./AlarmInfo.js'),
-    AlarmNotification: () => import('./AlarmNotification.js'),
-    AlarmServicePost: () => import('./AlarmServicePost.js'),
-    ArchiveRef: () => import('./ArchiveRef.js'),
-    ArchiveType: () => import('./ArchiveType.js'),
-    AuditLogEntry: () => import('./AuditLogEntry.js'),
-    AuditLogRef: () => import('./AuditLogRef.js'),
-    AuthenticatedDevice: () => import('./AuthenticatedDevice.js'),
-    Authentication: () => import('./Authentication.js'),
-    AutoLoginDataDelete: () => import('./AutoLoginDataDelete.js'),
-    AutoLoginDataGet: () => import('./AutoLoginDataGet.js'),
-    AutoLoginDataReturn: () => import('./AutoLoginDataReturn.js'),
-    AutoLoginPostReturn: () => import('./AutoLoginPostReturn.js'),
-    Blob: () => import('./Blob.js'),
-    BlobReferenceTokenWrapper: () => import('./BlobReferenceTokenWrapper.js'),
-    Booking: () => import('./Booking.js'),
-    BookingItem: () => import('./BookingItem.js'),
-    BookingServiceData: () => import('./BookingServiceData.js'),
-    BookingsRef: () => import('./BookingsRef.js'),
-    BootstrapFeature: () => import('./BootstrapFeature.js'),
-    Braintree3ds2Request: () => import('./Braintree3ds2Request.js'),
-    Braintree3ds2Response: () => import('./Braintree3ds2Response.js'),
-    BrandingDomainData: () => import('./BrandingDomainData.js'),
-    BrandingDomainDeleteData: () => import('./BrandingDomainDeleteData.js'),
-    BrandingDomainGetReturn: () => import('./BrandingDomainGetReturn.js'),
-    Bucket: () => import('./Bucket.js'),
-    BucketPermission: () => import('./BucketPermission.js'),
-    CalendarEventRef: () => import('./CalendarEventRef.js'),
-    CertificateInfo: () => import('./CertificateInfo.js'),
-    Challenge: () => import('./Challenge.js'),
-    ChangePasswordData: () => import('./ChangePasswordData.js'),
-    Chat: () => import('./Chat.js'),
-    CloseSessionServicePost: () => import('./CloseSessionServicePost.js'),
-    CreateCustomerServerPropertiesData: () => import('./CreateCustomerServerPropertiesData.js'),
-    CreateCustomerServerPropertiesReturn: () => import('./CreateCustomerServerPropertiesReturn.js'),
-    CreateGroupData: () => import('./CreateGroupData.js'),
-    CreateGroupListData: () => import('./CreateGroupListData.js'),
-    CreateSessionData: () => import('./CreateSessionData.js'),
-    CreateSessionReturn: () => import('./CreateSessionReturn.js'),
-    CreditCard: () => import('./CreditCard.js'),
-    CustomDomainCheckData: () => import('./CustomDomainCheckData.js'),
-    CustomDomainCheckReturn: () => import('./CustomDomainCheckReturn.js'),
-    CustomDomainData: () => import('./CustomDomainData.js'),
-    CustomDomainReturn: () => import('./CustomDomainReturn.js'),
-    Customer: () => import('./Customer.js'),
-    CustomerData: () => import('./CustomerData.js'),
-    CustomerInfo: () => import('./CustomerInfo.js'),
-    CustomerInfoReturn: () => import('./CustomerInfoReturn.js'),
-    CustomerProperties: () => import('./CustomerProperties.js'),
-    CustomerReturn: () => import('./CustomerReturn.js'),
-    CustomerServerProperties: () => import('./CustomerServerProperties.js'),
-    DebitServicePutData: () => import('./DebitServicePutData.js'),
-    DeleteCustomerData: () => import('./DeleteCustomerData.js'),
-    DnsRecord: () => import('./DnsRecord.js'),
-    DomainInfo: () => import('./DomainInfo.js'),
-    DomainMailAddressAvailabilityData: () => import('./DomainMailAddressAvailabilityData.js'),
-    DomainMailAddressAvailabilityReturn: () => import('./DomainMailAddressAvailabilityReturn.js'),
-    DomainsRef: () => import('./DomainsRef.js'),
-    EmailSenderListElement: () => import('./EmailSenderListElement.js'),
-    EntityEventBatch: () => import('./EntityEventBatch.js'),
-    EntityUpdate: () => import('./EntityUpdate.js'),
-    Exception: () => import('./Exception.js'),
-    ExternalPropertiesReturn: () => import('./ExternalPropertiesReturn.js'),
-    ExternalUserReference: () => import('./ExternalUserReference.js'),
-    Feature: () => import('./Feature.js'),
-    File: () => import('./File.js'),
-    GeneratedIdWrapper: () => import('./GeneratedIdWrapper.js'),
-    GiftCard: () => import('./GiftCard.js'),
-    GiftCardCreateData: () => import('./GiftCardCreateData.js'),
-    GiftCardCreateReturn: () => import('./GiftCardCreateReturn.js'),
-    GiftCardDeleteData: () => import('./GiftCardDeleteData.js'),
-    GiftCardGetReturn: () => import('./GiftCardGetReturn.js'),
-    GiftCardOption: () => import('./GiftCardOption.js'),
-    GiftCardRedeemData: () => import('./GiftCardRedeemData.js'),
-    GiftCardRedeemGetReturn: () => import('./GiftCardRedeemGetReturn.js'),
-    GiftCardsRef: () => import('./GiftCardsRef.js'),
-    Group: () => import('./Group.js'),
-    GroupInfo: () => import('./GroupInfo.js'),
-    GroupMember: () => import('./GroupMember.js'),
-    GroupMembership: () => import('./GroupMembership.js'),
-    GroupRoot: () => import('./GroupRoot.js'),
-    Invoice: () => import('./Invoice.js'),
-    InvoiceInfo: () => import('./InvoiceInfo.js'),
-    InvoiceItem: () => import('./InvoiceItem.js'),
-    KeyPair: () => import('./KeyPair.js'),
-    LocationServiceGetReturn: () => import('./LocationServiceGetReturn.js'),
-    Login: () => import('./Login.js'),
-    MailAddressAlias: () => import('./MailAddressAlias.js'),
-    MailAddressAliasServiceData: () => import('./MailAddressAliasServiceData.js'),
-    MailAddressAliasServiceDataDelete: () => import('./MailAddressAliasServiceDataDelete.js'),
-    MailAddressAliasServiceReturn: () => import('./MailAddressAliasServiceReturn.js'),
-    MailAddressAvailabilityData: () => import('./MailAddressAvailabilityData.js'),
-    MailAddressAvailabilityReturn: () => import('./MailAddressAvailabilityReturn.js'),
-    MailAddressToGroup: () => import('./MailAddressToGroup.js'),
-    MembershipAddData: () => import('./MembershipAddData.js'),
-    MembershipRemoveData: () => import('./MembershipRemoveData.js'),
-    MissedNotification: () => import('./MissedNotification.js'),
-    NotificationInfo: () => import('./NotificationInfo.js'),
-    NotificationMailTemplate: () => import('./NotificationMailTemplate.js'),
-    NotificationSessionKey: () => import('./NotificationSessionKey.js'),
-    OrderProcessingAgreement: () => import('./OrderProcessingAgreement.js'),
-    OtpChallenge: () => import('./OtpChallenge.js'),
-    PaymentDataServiceGetData: () => import('./PaymentDataServiceGetData.js'),
-    PaymentDataServiceGetReturn: () => import('./PaymentDataServiceGetReturn.js'),
-    PaymentDataServicePostData: () => import('./PaymentDataServicePostData.js'),
-    PaymentDataServicePutData: () => import('./PaymentDataServicePutData.js'),
-    PaymentDataServicePutReturn: () => import('./PaymentDataServicePutReturn.js'),
-    PaymentErrorInfo: () => import('./PaymentErrorInfo.js'),
-    PdfInvoiceServiceData: () => import('./PdfInvoiceServiceData.js'),
-    PdfInvoiceServiceReturn: () => import('./PdfInvoiceServiceReturn.js'),
-    Permission: () => import('./Permission.js'),
-    PhoneNumber: () => import('./PhoneNumber.js'),
-    PlanPrices: () => import('./PlanPrices.js'),
-    PremiumFeatureData: () => import('./PremiumFeatureData.js'),
-    PremiumFeatureReturn: () => import('./PremiumFeatureReturn.js'),
-    PriceData: () => import('./PriceData.js'),
-    PriceItemData: () => import('./PriceItemData.js'),
-    PriceRequestData: () => import('./PriceRequestData.js'),
-    PriceServiceData: () => import('./PriceServiceData.js'),
-    PriceServiceReturn: () => import('./PriceServiceReturn.js'),
-    PublicKeyData: () => import('./PublicKeyData.js'),
-    PublicKeyReturn: () => import('./PublicKeyReturn.js'),
-    PushIdentifier: () => import('./PushIdentifier.js'),
-    PushIdentifierList: () => import('./PushIdentifierList.js'),
-    ReceivedGroupInvitation: () => import('./ReceivedGroupInvitation.js'),
-    RecoverCode: () => import('./RecoverCode.js'),
-    RegistrationCaptchaServiceData: () => import('./RegistrationCaptchaServiceData.js'),
-    RegistrationCaptchaServiceGetData: () => import('./RegistrationCaptchaServiceGetData.js'),
-    RegistrationCaptchaServiceReturn: () => import('./RegistrationCaptchaServiceReturn.js'),
-    RegistrationConfigReturn: () => import('./RegistrationConfigReturn.js'),
-    RegistrationReturn: () => import('./RegistrationReturn.js'),
-    RegistrationServiceData: () => import('./RegistrationServiceData.js'),
-    RejectedSender: () => import('./RejectedSender.js'),
-    RejectedSendersRef: () => import('./RejectedSendersRef.js'),
-    RepeatRule: () => import('./RepeatRule.js'),
-    ResetFactorsDeleteData: () => import('./ResetFactorsDeleteData.js'),
-    ResetPasswordData: () => import('./ResetPasswordData.js'),
-    RootInstance: () => import('./RootInstance.js'),
-    SaltData: () => import('./SaltData.js'),
-    SaltReturn: () => import('./SaltReturn.js'),
-    SecondFactor: () => import('./SecondFactor.js'),
-    SecondFactorAuthAllowedReturn: () => import('./SecondFactorAuthAllowedReturn.js'),
-    SecondFactorAuthData: () => import('./SecondFactorAuthData.js'),
-    SecondFactorAuthDeleteData: () => import('./SecondFactorAuthDeleteData.js'),
-    SecondFactorAuthGetData: () => import('./SecondFactorAuthGetData.js'),
-    SecondFactorAuthGetReturn: () => import('./SecondFactorAuthGetReturn.js'),
-    SecondFactorAuthentication: () => import('./SecondFactorAuthentication.js'),
-    SendRegistrationCodeData: () => import('./SendRegistrationCodeData.js'),
-    SendRegistrationCodeReturn: () => import('./SendRegistrationCodeReturn.js'),
-    SentGroupInvitation: () => import('./SentGroupInvitation.js'),
-    Session: () => import('./Session.js'),
-    SignOrderProcessingAgreementData: () => import('./SignOrderProcessingAgreementData.js'),
-    SseConnectData: () => import('./SseConnectData.js'),
-    StringConfigValue: () => import('./StringConfigValue.js'),
-    StringWrapper: () => import('./StringWrapper.js'),
-    SwitchAccountTypeData: () => import('./SwitchAccountTypeData.js'),
-    SystemKeysReturn: () => import('./SystemKeysReturn.js'),
-    TakeOverDeletedAddressData: () => import('./TakeOverDeletedAddressData.js'),
-    TypeInfo: () => import('./TypeInfo.js'),
-    U2fChallenge: () => import('./U2fChallenge.js'),
-    U2fKey: () => import('./U2fKey.js'),
-    U2fRegisteredDevice: () => import('./U2fRegisteredDevice.js'),
-    U2fResponseData: () => import('./U2fResponseData.js'),
-    UpdateAdminshipData: () => import('./UpdateAdminshipData.js'),
-    UpdatePermissionKeyData: () => import('./UpdatePermissionKeyData.js'),
-    UpgradePriceServiceData: () => import('./UpgradePriceServiceData.js'),
-    UpgradePriceServiceReturn: () => import('./UpgradePriceServiceReturn.js'),
-    User: () => import('./User.js'),
-    UserAlarmInfo: () => import('./UserAlarmInfo.js'),
-    UserAlarmInfoListType: () => import('./UserAlarmInfoListType.js'),
-    UserAreaGroups: () => import('./UserAreaGroups.js'),
-    UserAuthentication: () => import('./UserAuthentication.js'),
-    UserData: () => import('./UserData.js'),
-    UserDataDelete: () => import('./UserDataDelete.js'),
-    UserExternalAuthInfo: () => import('./UserExternalAuthInfo.js'),
-    UserGroupRoot: () => import('./UserGroupRoot.js'),
-    UserIdData: () => import('./UserIdData.js'),
-    UserIdReturn: () => import('./UserIdReturn.js'),
-    UserReturn: () => import('./UserReturn.js'),
-    VariableExternalAuthInfo: () => import('./VariableExternalAuthInfo.js'),
-    VerifyRegistrationCodeData: () => import('./VerifyRegistrationCodeData.js'),
-    Version: () => import('./Version.js'),
-    VersionData: () => import('./VersionData.js'),
-    VersionInfo: () => import('./VersionInfo.js'),
-    VersionReturn: () => import('./VersionReturn.js'),
-    WebauthnResponseData: () => import('./WebauthnResponseData.js'),
-    WebsocketCounterData: () => import('./WebsocketCounterData.js'),
-    WebsocketCounterValue: () => import('./WebsocketCounterValue.js'),
-    WebsocketEntityData: () => import('./WebsocketEntityData.js'),
-    WebsocketLeaderStatus: () => import('./WebsocketLeaderStatus.js'),
-    WhitelabelChild: () => import('./WhitelabelChild.js'),
-    WhitelabelChildrenRef: () => import('./WhitelabelChildrenRef.js'),
-    WhitelabelConfig: () => import('./WhitelabelConfig.js'),
-    WhitelabelParent: () => import('./WhitelabelParent.js')
-}
-export default map
