+++ src/api/entities/usage/ModelInfo.ts
@@ -1,0 +1,5 @@
+const modelInfo = {
+	version: 1,
+}
+		
+export default modelInfo
