--- src/api/entities/usage/usageModelMapDebug.ts
@@ -1,10 +1,0 @@
-const map = {    "UsageTestAssignment": () => Promise.resolve({"_TypeModel":{"name":"UsageTestAssignment","since":1,"type":"AGGREGATED_TYPE","id":56,"rootId":"BXVzYWdlADg","versioned":false,"encrypted":false,"values":{"_id":{"id":57,"type":"CustomId","cardinality":"One","final":true,"encrypted":false},"name":{"id":59,"type":"String","cardinality":"One","final":false,"encrypted":false},"sendPings":{"id":61,"type":"Boolean","cardinality":"One","final":false,"encrypted":false},"testId":{"id":58,"type":"GeneratedId","cardinality":"One","final":true,"encrypted":false},"variant":{"id":60,"type":"Number","cardinality":"ZeroOrOne","final":true,"encrypted":false}},"associations":{"stages":{"id":62,"type":"AGGREGATION","cardinality":"Any","final":false,"refType":"UsageTestStage","dependency":null}},"app":"usage","version":"1"}}),
-    "UsageTestAssignmentIn": () => Promise.resolve({"_TypeModel":{"name":"UsageTestAssignmentIn","since":1,"type":"DATA_TRANSFER_TYPE","id":53,"rootId":"BXVzYWdlADU","versioned":false,"encrypted":false,"values":{"_format":{"id":54,"type":"Number","cardinality":"One","final":false,"encrypted":false},"testDeviceId":{"id":55,"type":"GeneratedId","cardinality":"ZeroOrOne","final":false,"encrypted":false}},"associations":{},"app":"usage","version":"1"}}),
-    "UsageTestAssignmentOut": () => Promise.resolve({"_TypeModel":{"name":"UsageTestAssignmentOut","since":1,"type":"DATA_TRANSFER_TYPE","id":63,"rootId":"BXVzYWdlAD8","versioned":false,"encrypted":false,"values":{"_format":{"id":64,"type":"Number","cardinality":"One","final":false,"encrypted":false},"testDeviceId":{"id":65,"type":"GeneratedId","cardinality":"One","final":false,"encrypted":false}},"associations":{"assignments":{"id":66,"type":"AGGREGATION","cardinality":"Any","final":false,"refType":"UsageTestAssignment","dependency":null}},"app":"usage","version":"1"}}),
-    "UsageTestMetricConfig": () => Promise.resolve({"_TypeModel":{"name":"UsageTestMetricConfig","since":1,"type":"AGGREGATED_TYPE","id":12,"rootId":"BXVzYWdlAAw","versioned":false,"encrypted":false,"values":{"_id":{"id":13,"type":"CustomId","cardinality":"One","final":true,"encrypted":false},"name":{"id":14,"type":"String","cardinality":"One","final":true,"encrypted":false},"type":{"id":15,"type":"Number","cardinality":"One","final":true,"encrypted":false}},"associations":{"configValues":{"id":16,"type":"AGGREGATION","cardinality":"Any","final":false,"refType":"UsageTestMetricConfigValue","dependency":null}},"app":"usage","version":"1"}}),
-    "UsageTestMetricConfigValue": () => Promise.resolve({"_TypeModel":{"name":"UsageTestMetricConfigValue","since":1,"type":"AGGREGATED_TYPE","id":8,"rootId":"BXVzYWdlAAg","versioned":false,"encrypted":false,"values":{"_id":{"id":9,"type":"CustomId","cardinality":"One","final":true,"encrypted":false},"key":{"id":10,"type":"String","cardinality":"One","final":false,"encrypted":false},"value":{"id":11,"type":"String","cardinality":"One","final":false,"encrypted":false}},"associations":{},"app":"usage","version":"1"}}),
-    "UsageTestMetricData": () => Promise.resolve({"_TypeModel":{"name":"UsageTestMetricData","since":1,"type":"AGGREGATED_TYPE","id":17,"rootId":"BXVzYWdlABE","versioned":false,"encrypted":false,"values":{"_id":{"id":18,"type":"CustomId","cardinality":"One","final":true,"encrypted":false},"name":{"id":19,"type":"String","cardinality":"One","final":true,"encrypted":false},"value":{"id":20,"type":"String","cardinality":"One","final":true,"encrypted":false}},"associations":{},"app":"usage","version":"1"}}),
-    "UsageTestParticipationIn": () => Promise.resolve({"_TypeModel":{"name":"UsageTestParticipationIn","since":1,"type":"DATA_TRANSFER_TYPE","id":80,"rootId":"BXVzYWdlAFA","versioned":false,"encrypted":false,"values":{"_format":{"id":81,"type":"Number","cardinality":"One","final":false,"encrypted":false},"stage":{"id":83,"type":"Number","cardinality":"One","final":false,"encrypted":false},"testDeviceId":{"id":84,"type":"GeneratedId","cardinality":"One","final":false,"encrypted":false},"testId":{"id":82,"type":"GeneratedId","cardinality":"One","final":false,"encrypted":false}},"associations":{"metrics":{"id":85,"type":"AGGREGATION","cardinality":"Any","final":false,"refType":"UsageTestMetricData","dependency":null}},"app":"usage","version":"1"}}),
-    "UsageTestStage": () => Promise.resolve({"_TypeModel":{"name":"UsageTestStage","since":1,"type":"AGGREGATED_TYPE","id":35,"rootId":"BXVzYWdlACM","versioned":false,"encrypted":false,"values":{"_id":{"id":36,"type":"CustomId","cardinality":"One","final":true,"encrypted":false},"name":{"id":37,"type":"String","cardinality":"One","final":false,"encrypted":false}},"associations":{"metrics":{"id":38,"type":"AGGREGATION","cardinality":"Any","final":false,"refType":"UsageTestMetricConfig","dependency":null}},"app":"usage","version":"1"}}),
-}
-export default map
